// creditgo copy/src/models/counterparty-entity.ts
export interface CounterpartyEntity {
    id: number;
    name: string;
    parentName?: any;
    jurisdiction: string;
    isOnAlert: boolean;
    alertNotes?: any;
    industryClass: string;
    appraisalCompletionDate: Date;
    nextAppraisalDueDate: Date;
    gcrAnalyst: string;
    strategicCreditAnalyst: string;
    countryOfRisk: string;
    rbuOwner: string;
    fiscalYearEndDate?: any;
    isMonitored: boolean;
    isWatched: boolean;
    watchComments?: any;
    canProvideCollateral: boolean;
    requiresBankLoi: boolean;
    permitsEpfTrades: boolean;
    bpRating: string;
    bpConfidenceLevel: string;
    portfolioTag?: any;
    bpOutlook: string;
    riskLevel: string;
    sppRating?: any;
    moodyRating?: any;
    fitchRating?: any;
    issuerSppRating?: any;
    issuerMoodyRating?: any;
    isuerFitchRating?: any;
    sppOutlook?: any;
    moodyOutlook?: any;
    fitchOutlook?: any;
    bpRatingInferred: string;
    industryInferred: string;
    inferredBpRatingCeId: number;
    inferredIndustryCeId: number;
    isPfeCounterparty: boolean;
    creditLimit_0_6M: number;
    creditLimit_7_12M: number;
    creditLimit_13_24M: number;
    creditLimit_25M_Plus: number;
    creditLimitCurrency?: any;
    pfeComments?: any;
}